+++ Timelapse2/Timelapse/Util/Throttles.cs
@@ -37,7 +37,7 @@ namespace Timelapse.Util
 
             this.DesiredImageRendersPerSecond = rendersPerSecond;
             this.DesiredIntervalBetweenRenders = TimeSpan.FromSeconds(1.0 / rendersPerSecond);
-            this.RepeatedKeyAcceptanceInterval = (int)(((double)SystemParameters.KeyboardSpeed + 0.5 * rendersPerSecond) / rendersPerSecond);
+            this.RepeatedKeyAcceptanceInterval = (int)((SystemParameters.KeyboardSpeed + 0.5 * rendersPerSecond) / rendersPerSecond);
         }
     }
 }
